feat: migrate data module to KMP
