+++ build.gradle.kts
@@ -5,10 +5,10 @@ plugins {
     alias(libs.plugins.android.library) apply false
     alias(libs.plugins.kotlin.jvm) apply false
     alias(libs.plugins.ksp) apply false
-    alias(libs.plugins.hilt) apply false
     alias(libs.plugins.kotlin.multiplatform) apply false
     alias(libs.plugins.android.kotlin.multiplatform.library) apply false
     alias(libs.plugins.android.lint) apply false
+    alias(libs.plugins.androidx.room) apply false
 }
 
 buildscript {
@@ -24,7 +24,6 @@ buildscript {
     }
 
     dependencies {
-        classpath(libs.hilt.android.gradle)
         classpath(libs.spotless.gradle)
     }
 }
