+++ data/src/commonMain/kotlin/com/mindeck/data/clock/ClockRepositoryImpl.kt
@@ -1,0 +1,8 @@
+package com.mindeck.data.clock
+
+import com.mindeck.domain.service.ClockRepository
+import kotlin.time.Clock
+
+class ClockRepositoryImpl : ClockRepository {
+    override fun now(): Long = Clock.System.now().toEpochMilliseconds()
+}
