+++ data/src/commonMain/kotlin/com/mindeck/data/entities/CardWithDeckEntity.kt
@@ -1,0 +1,13 @@
+package com.mindeck.data.entities
+
+import androidx.room.Embedded
+import androidx.room.Relation
+
+data class CardWithDeckEntity(
+    @Embedded val card: CardEntity,
+    @Relation(
+        parentColumn = "deck_id",
+        entityColumn = "deck_id",
+    )
+    val deck: DeckEntity,
+)
