+++ data/src/commonMain/kotlin/com/mindeck/data/entities/DeckEntity.kt
@@ -1,0 +1,22 @@
+package com.mindeck.data.entities
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.Index
+import androidx.room.PrimaryKey
+
+@Entity(
+    tableName = "deck",
+    indices = [
+        Index(
+            value = ["deck_name"],
+            unique = true,
+        ),
+    ],
+)
+data class DeckEntity(
+    @PrimaryKey(autoGenerate = true)
+    @ColumnInfo(name = "deck_id")
+    val deckId: Int = 0,
+    @ColumnInfo(name = "deck_name") val deckName: String,
+)
