--- data/src/main/java/com/mindeck/data/clock/ClockRepositoryImpl.kt
@@ -1,7 +1,0 @@
-package com.mindeck.data.clock
-
-import com.mindeck.domain.service.ClockRepository
-
-class ClockRepositoryImpl : ClockRepository {
-    override fun now(): Long = System.currentTimeMillis()
-}
