--- data/src/main/java/com/mindeck/data/database/entities/CardWithDeckEntity.kt
@@ -1,13 +1,0 @@
-package com.mindeck.data.database.entities
-
-import androidx.room.Embedded
-import androidx.room.Relation
-
-data class CardWithDeckEntity(
-    @Embedded val card: CardEntity,
-    @Relation(
-        parentColumn = "deck_id",
-        entityColumn = "deck_id",
-    )
-    val deck: DeckEntity,
-)
